requirements.txt
streamlit==1.32.0
openai==1.12.0
python-dotenv==1.0.1
pandas==2.2.1
plotly==5.19.0
tiktoken==0.6.0
python-json-logger==2.0.7 